+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Remove extracted cops in `Capybara`, `FactoryBot` and `Rails` departments. ([@ydah])
 - Remove `RuboCop::RSpec::Language::NodePattern`. ([@ydah])
+- Remove `RSpec/FilePath` cop. ([@ydah])
 
 ## 2.29.2 (2024-05-02)